Updated environment scripts as per opam init
Note: as for right now, opam init detected my shell and location of
default init script (fish and ~/.config/fish/config.fish respectively)
correctly, and offered to append command to source a fish specific setup
script, but it failed to provide that script. I left it as is, in the
hope that this script will be provided in later version.
Meanwhile, I can use `eval (opam config env)` to add environment
variables as needed.